Module tor_llcrypto::d

Digests and XOFs used to implement the Tor protocol.

In various places, for legacy reasons, Tor uses SHA1, SHA2, SHA3, and SHAKE. We re-export them all here, in forms implementing the the digest::Digest traits.

Other code should access these digests via the traits in the digest crate.

Type Definitions

Sha1

SHA-1 hasher state.

Sha3_256

SHA-3-256 hasher state.

Sha256

SHA-256 hasher.

Sha512

SHA-512 hasher.

Shake128

SHAKE128 hasher state.

Shake256

SHAKE256 hasher state.